+++ test/Origin.cache.js
@@ -1,6 +1,6 @@
 const test = require("mocha-sinon-chai");
 
-const { Origin } = require("../src/Origin");
+const { Origin, sources } = require("../src/Origin");
 
 test.describe("Origin cache", () => {
   let sandbox;
@@ -37,6 +37,7 @@ test.describe("Origin cache", () => {
 
   test.afterEach(() => {
     sandbox.restore();
+    sources.clear();
   });
 
   test.describe("without query", () => {