+++ test/Origin.events.js
@@ -1,6 +1,6 @@
 const test = require("mocha-sinon-chai");
 
-const { Origin } = require("../src/Origin");
+const { Origin, sources } = require("../src/Origin");
 
 test.describe("Origin events", () => {
   let sandbox;
@@ -28,6 +28,7 @@ test.describe("Origin events", () => {
 
   test.afterEach(() => {
     sandbox.restore();
+    sources.clear();
   });
 
   test.describe("Without query", () => {