+++ test/Origin.id.js
@@ -1,6 +1,6 @@
 const test = require("mocha-sinon-chai");
 
-const { Origin } = require("../src/Origin");
+const { Origin, sources } = require("../src/Origin");
 const helpers = require("../src/helpers");
 
 test.describe("Origin id", () => {
@@ -18,6 +18,7 @@ test.describe("Origin id", () => {
 
   test.afterEach(() => {
     sandbox.restore();
+    sources.clear();
   });
 
   test.describe("Without query", () => {