+++ test/Origin.loading.js
@@ -1,6 +1,6 @@
 const test = require("mocha-sinon-chai");
 
-const { Origin } = require("../src/Origin");
+const { Origin, sources } = require("../src/Origin");
 
 test.describe("Origin loading", () => {
   let sandbox;
@@ -23,6 +23,7 @@ test.describe("Origin loading", () => {
 
   test.afterEach(() => {
     sandbox.restore();
+    sources.clear();
   });
 
   test.describe("using getter", () => {