+++ test/Origin.queries.js
@@ -1,6 +1,6 @@
 const test = require("mocha-sinon-chai");
 
-const { Origin } = require("../src/Origin");
+const { Origin, sources } = require("../src/Origin");
 
 test.describe("Origin queries", () => {
   const FOO_QUERY = { foo: "foo", foo2: "foo-2" };
@@ -42,6 +42,7 @@ test.describe("Origin queries", () => {
 
   test.afterEach(() => {
     sandbox.restore();
+    sources.clear();
   });
 
   test.describe("when executing methods", () => {